// packages/shards-meteor-process-manager-server/src/types.ts
import type { DefineProcessesParams } from '@moxb/shards-meteor-process-manager-core';

export type RequestType = 'defineProcesses' | 'launchProcess' | 'stopProcess';

export interface ProcessRequestCore {
    _id: string;
    scopeId: string;
    type: RequestType;
    requestedAt: Date;
    takenByWorker?: string;
}

export interface DefineProcessesRequest extends ProcessRequestCore {
    type: 'defineProcesses';
    params: DefineProcessesParams;
}

export interface LaunchProcessRequest extends ProcessRequestCore {
    type: 'launchProcess';
    processId: string;
}

export interface StopProcessRequest extends ProcessRequestCore {
    type: 'stopProcess';
    processId: string;
}

export type ProcessRequest = DefineProcessesRequest | LaunchProcessRequest | StopProcessRequest;